expの小数対応

+++ src/main/scala/com/ryoserver/Level/Player/getPlayerData.scala
@@ -8,15 +8,15 @@ import org.bukkit.entity.Player
 
 class getPlayerData(ryoServerAssist: RyoServerAssist) {
 
-  def getPlayerExp(p: Player):Int = {
+  def getPlayerExp(p: Player):Double = {
     new createData(ryoServerAssist).createPlayerData(p)
     val sql = new SQL(ryoServerAssist)
     val rs = sql.executeQuery(s"SELECT EXP,Level FROM Players WHERE UUID='${p.getUniqueId.toString}'")
-    if (rs.next()) return rs.getInt("EXP")
+    if (rs.next()) return rs.getDouble("EXP")
     sql.close()
     0
   }
 
-  def getPlayerLevel(p: Player):Int = new CalLv(ryoServerAssist).getLevel(getPlayerExp(p))
+  def getPlayerLevel(p: Player):Int = new CalLv(ryoServerAssist).getLevel(getPlayerExp(p).toInt)
 
 }
